requirements.txt
nltk==3.6.2
amqp==5.0.2
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==8.0.1
pytesseract==0.3.6
python-dotenv==0.15.0
nltk==3.6.2
imutils==0.5.3
pdf2image==1.14.0
Flask==1.1.2